Create  SampleGlideModule.java
This is reqiured to use the Glide in our app.Use Glide.with(this) everywhere
Glide takes any among the Activity, Context, or Fragment. The benefit of passing in the Fragment or Activity? Image loading, its memory management, and the pausing and resuming of GIFs are then in step with the lifecycle of the activity. So Glide attaches its life cycle with Activity, Context, or Fragment and onPause(), onStop(), onRestart() etc it behaves accordingly.